Break out tab logic into its own file
