+++ libs/extension-messaging/src/constants.ts
@@ -1,0 +1,5 @@
+// An identifier that tells us which version of the messaging service we're using,
+// to ensure we don't try to process a message not intended for us.
+const version = 2.5;
+
+export { version };
